+++ chain/api/src/service.rs
@@ -56,7 +56,7 @@ pub trait ReadableChainService {
         &self,
         start_number: BlockNumber,
         reverse: bool,
-        max_size: usize,
+        max_size: u64,
     ) -> Result<Vec<HashValue>>;
 }
 
@@ -113,7 +113,7 @@ pub trait ChainAsyncService:
         &self,
         start_number: BlockNumber,
         reverse: bool,
-        max_size: usize,
+        max_size: u64,
     ) -> Result<Vec<HashValue>>;
 }
 
@@ -367,7 +367,7 @@ where
         &self,
         start_number: BlockNumber,
         reverse: bool,
-        max_size: usize,
+        max_size: u64,
     ) -> Result<Vec<HashValue>> {
         let response = self
             .send(ChainRequest::GetBlockIds {
